Move the divider dragging logic back to the dedicated diider component
